openwrt: fix gcc-{ar,nm,ranlib} symlinks
These should absoluetly not point to the gcc wrapper, they are thin
wrappers around the binutils tools that are necessary to make LTO work.

This fix link failures in apk, where Meson would (I assume) see a broken
x86_64-unknown-linux-gnu-gcc-ar (or a missing gcc-ar), and fall back to
the unwrapped ar.